--- app/controllers/organizations_controller.rb
@@ -1,57 +1,0 @@
-class OrganizationsController < ApplicationController
-  before_action :set_organization, only: %i[ show update destroy users]
-
-  # GET /organizations
-  def index
-    @organizations = Organization.all
-
-    render json: @organizations
-  end
-
-  # GET /organizations/1
-  def show
-    render json: @organization
-  end
-
-  # POST /organizations
-  def create
-    @organization = Organization.new(organization_params)
-
-    if @organization.save
-      render json: @organization, status: :created, location: @organization
-    else
-      render json: @organization.errors, status: :unprocessable_entity
-    end
-  end
-
-  # PATCH/PUT /organizations/1
-  def update
-    if @organization.update(organization_params)
-      render json: @organization
-    else
-      render json: @organization.errors, status: :unprocessable_entity
-    end
-  end
-
-  # DELETE /organizations/1
-  def destroy
-    @organization.destroy!
-  end
-
-   # GET /organizations/:id/users
-   def users
-    @users = @organization.users
-    render json: @users
-  end
-
-  private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_organization
-      @organization = Organization.find(params[:id])
-    end
-
-    # Only allow a list of trusted parameters through.
-    def organization_params
-      params.require(:organization).permit(:name, :address, :email, :web_address, :subdomain)
-    end
-end
