+++ app/serializers/team_serializer.rb
@@ -1,0 +1,6 @@
+# app/serializers/team_serializer.rb
+class TeamSerializer
+  include JSONAPI::Serializer
+
+  attributes :id, :name, :created_at, :updated_at
+end
